Adds opacity to voting center and fixes jump
